requirements.txt
selenium~=4.33.0
webdriver-manager
pytest~=8.4.1
allure-pytest
Faker~=37.4.0